first-time link opening done by visit() instead of typst-preview cli
- move the definition of `is_xxx_system()`, `is_xxx_arch()` and `visit()` to utils.lua
- link opening would be handled by lua function in the plugin instead of typst-preview CLI when calling command `TypstPreview` the first time
- print debug info in visit()
- change the open command in WSL to `/mnt/c/Windows/explorer.exe`
