+++ lua/typst-preview/utils.lua
@@ -1,6 +1,70 @@
 local config = require 'typst-preview.config'
 local M = {}
 
+---check if the host system is windows
+function M.is_windows()
+  return vim.loop.os_uname().sysname == 'Windows_NT'
+end
+
+---check if the host system is macos
+function M.is_macos()
+  return vim.loop.os_uname().sysname == 'Darwin'
+end
+
+---check if the host system is linux
+function M.is_linux()
+  return vim.loop.os_uname().sysname == 'Linux'
+end
+
+---check if the host system is wsl
+function M.is_wsl()
+  return M.is_linux() and vim.loop.os_uname().release:lower():find 'microsoft'
+end
+
+-- Stolen from mason.nvim
+
+---check if the host arch is x64
+function M.is_x64()
+  local machine = vim.loop.os_uname().machine
+  return machine == 'x86_64' or machine == 'x64'
+end
+
+---check if the host arch is arm64
+function M.is_arm64()
+  local machine = vim.loop.os_uname().machine
+  return machine == 'aarch64'
+    or machine == 'aarch64_be'
+    or machine == 'armv8b'
+    or machine == 'armv8l'
+    or machine == 'arm64'
+end
+
+local open_cmd
+if M.is_macos() then
+  open_cmd = 'open'
+elseif M.is_windows() then
+  open_cmd = 'explorer.exe'
+elseif M.is_wsl() then
+  open_cmd = '/mnt/c/Windows/explorer.exe'
+else
+  open_cmd = 'xdg-open'
+end
+
+---Open link in browser (platform agnostic)
+---@param link string
+function M.visit(link)
+  local cmd = string.format('%s http://%s', open_cmd, link)
+  M.debug(cmd)
+  vim.fn.jobstart(cmd, {
+    on_stderr = function(_, data)
+      local msg = table.concat(data or {}, '\n')
+      if msg ~= '' then
+        print('typst-preview opening link failed: ' .. msg)
+      end
+    end,
+  })
+end
+
 ---check if a file exist
 ---@param path string
 function M.file_exist(path)
